Chained useFrequency to be a float

+++ core/src/com/unciv/models/UnitAction.kt
@@ -19,7 +19,7 @@ open class UnitAction(
      * 100 is very frequent, 50 is somewhat frequent, less than 25 is press one time for multi-turn movement. 
      * A Rare case is > 100 if a button is something like add in capital, promote or something,
      * we need to inform the player that taking the action is an option. */
-    val useFrequency: Int,
+    val useFrequency: Float,
     val title: String = type.value,
     val isCurrentAction: Boolean = false,
     val uncivSound: UncivSound = type.uncivSound,
@@ -79,7 +79,7 @@ class UpgradeUnitAction(
     val goldCostOfUpgrade: Int,
     val newResourceRequirements: Counter<String>,
     action: (() -> Unit)?
-) : UnitAction(UnitActionType.Upgrade, 120, title, action = action)
+) : UnitAction(UnitActionType.Upgrade, 120f, title, action = action)
 
 /** Unit Actions - generic enum with static properties
  *
